added users and http reqs
