--- .idea/libraries/android_android_24.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="android-android-24">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/AppData/Local/Android/sdk/platforms/android-24/android.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
