+++ app/src/main/java/com/example/stephen/housecall/UserControl.java
@@ -1,6 +1,6 @@
 package com.example.stephen.housecall;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
+import com.example.stephen.housecall.models.HTTP_REQUEST;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
@@ -11,22 +11,22 @@ import java.io.IOException;
 public class UserControl {
 
 
-    public static String createPatient(Patient patient){
-
-
-
-        ObjectMapper om = new ObjectMapper();
-        try{
-            String json = om.writeValueAsString(patient);
-            String msg =  HTTP_REQUEST.excutePost("https://92826306-0970-45ee-9fe1-aaf4416cf421-bluemix.cloudant.com/patients/_all_docs", json);
-            return msg;
-
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
-
-        return "Failed";
-    }
+//    public static String createPatient(Patient patient){
+//
+//
+//
+//        ObjectMapper om = new ObjectMapper();
+//        try{
+//            String json = om.writeValueAsString(patient);
+//            String msg =  HTTP_REQUEST.excutePost("https://92826306-0970-45ee-9fe1-aaf4416cf421-bluemix.cloudant.com/patients/_all_docs", json);
+//            return msg;
+//
+//        } catch (JsonProcessingException e) {
+//            e.printStackTrace();
+//        }
+//
+//        return "Failed";
+//    }
 
     public Patient findPatient(Patient patient) throws IOException {
 
@@ -41,20 +41,20 @@ public class UserControl {
 
     }
 
-    public String createDoc(Doctor doc){
-
-        ObjectMapper om = new ObjectMapper();
-        try{
-            String json = om.writeValueAsString(doc);
-            String msg =  HTTP_REQUEST.excutePost("IBMurl", json);
-            return msg;
-
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
-
-        return "Failed";
-    }
+//    public String createDoc(Doctor doc){
+//
+//        ObjectMapper om = new ObjectMapper();
+//        try{
+//            String json = om.writeValueAsString(doc);
+//            String msg =  HTTP_REQUEST.excutePost("IBMurl", json);
+//            return msg;
+//
+//        } catch (JsonProcessingException e) {
+//            e.printStackTrace();
+//        }
+//
+//        return "Failed";
+//    }
 
     public Doctor findDoc(Doctor doc) throws IOException {
